Move priorityQueue to a separate module
